operator: extract config layering
